Sync project published version with page published version
