+++ app/controllers/projects_controller.rb
@@ -1,12 +1,12 @@
 class ProjectsController < ApplicationController
   def show
-    @project = Project.find(params[:id]).revisions.last # TODO published revision
+    @project = Project.find(params[:id]).published_revision
     @rating_types_by_phase = RatingType.all.group_by(&:rating_phase)
     @ratings_by_type = @project.ratings.index_by(&:rating_type)
   end
 
   def index
     @rating_types_by_phase = RatingType.all.group_by(&:rating_phase)
-    @projects = Project.all.map { |p| p.revisions.last } # TODO published revision
+    @projects = Project.published.map { |p| p.published_revision } # TODO published revision
   end
 end
