+++ db/migrate/20171031125640_add_published_revision_to_project.rb
@@ -1,0 +1,5 @@
+class AddPublishedRevisionToProject < ActiveRecord::Migration[5.1]
+  def change
+    add_reference :projects, :published_revision, foreign_key: { to_table: :project_revisions }
+  end
+end
