--- spec/models/project_spec.rb
@@ -1,23 +1,0 @@
-# == Schema Information
-#
-# Table name: projects
-#
-#  id         :integer          not null, primary key
-#  page_id    :integer          not null
-#  created_at :datetime         not null
-#  updated_at :datetime         not null
-#
-# Indexes
-#
-#  index_projects_on_page_id  (page_id)
-#
-# Foreign Keys
-#
-#  fk_rails_...  (page_id => pages.id)
-#
-
-require 'rails_helper'
-
-RSpec.describe Project, type: :model do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
