--- spec/models/rating_phase_spec.rb
@@ -1,15 +1,0 @@
-# == Schema Information
-#
-# Table name: rating_phases
-#
-#  id         :integer          not null, primary key
-#  name       :string           not null
-#  created_at :datetime         not null
-#  updated_at :datetime         not null
-#
-
-require 'rails_helper'
-
-RSpec.describe RatingPhase, type: :model do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
